refactor(celery): delete empty shims and update imports to single source
- Delete `celery_reliability.py` and `celery_registry.py` shims (94 + 25 lines)
  — all code now lives in `celery.py` (single source, 1221 lines). Imports
  updated: `tests/unit/celery/test_reliability_system.py`,
  `tests/unit/test_circuit_breaker.py`, `src/tasks/*`,
  `src/app/shared/outbox/relay.py`, `src/app/connections/__init__.py` now
  `from app.connections.celery import ...`.
- `test_reliability_system.py` now tests the new file directly (was importing
  from shim, now from `celery`). `test_typed_dispatch.py` probe string updated.
- `celery_task_names.py` stays as single definition site (imported by celery.py)
  so `TASK_DECLARING_MODULES` patching still works with `real_celery` fixture.
- No dataclass left in `celery.py` — `RateLimitResult` is `BaseModel(frozen=True)`
  per clarification.

Related to #44
