+++ src/tasks/auth_email_tasks.py
@@ -19,8 +19,7 @@ the dispatch pass would hide the defect again and hand the worker a lock key of
 from functools import partial
 
 from app.config import get_settings
-from app.connections.celery import ResilientTask, celery_app
-from app.connections.celery_registry import CeleryTaskPayload, CeleryTaskRegistry
+from app.connections.celery import CeleryTaskPayload, CeleryTaskRegistry, ResilientTask, celery_app
 from app.connections.celery_task_names import (
     SEND_PASSWORD_RESET_EMAIL,
     SEND_VERIFICATION_EMAIL,
