+++ src/tasks/credit_tasks.py
@@ -7,8 +7,7 @@ import asyncio
 from returns.result import Failure
 from sqlalchemy import select
 
-from app.connections.celery import ResilientTask, celery_app
-from app.connections.celery_registry import CeleryTaskRegistry, NoKwargsPayload
+from app.connections.celery import CeleryTaskRegistry, NoKwargsPayload, ResilientTask, celery_app
 from app.connections.celery_task_names import CREDITS_EXPIRE, CREDITS_RECONCILE
 from app.connections.postgres import init_db
 from app.features.audit.model import AuditLog
